+++ src/platform/apple/methods/error/set.rs
@@ -2,10 +2,10 @@ use crate::error::Error;
 
 impl Error {
     pub fn set(self) {
-        Self::set_raw(self as i32);
+        Self::set_raw(self as i64);
     }
 
-    pub fn set_raw(err: i32) {
-        unsafe { *libc::__error() = err };
+    pub fn set_raw(err: i64) {
+        unsafe { *libc::__error() = err as i32 };
     }
 }
