+++ src/platform/apple/methods/file/fs/file_metadata/id.rs
@@ -1,11 +1,11 @@
 use crate::file::FileMetadata;
 
 impl FileMetadata {
-    pub fn uid(&self) -> u32 {
-        return self.metadata.st_uid;
+    pub fn uid(&self) -> u64 {
+        return self.metadata.st_uid as u64;
     }
 
-    pub fn gid(&self) -> u32 {
-        return self.metadata.st_gid;
+    pub fn gid(&self) -> u64 {
+        return self.metadata.st_gid as u64;
     }
 }
