+++ src/platform/apple/methods/file/io/tell.rs
@@ -2,7 +2,7 @@ use crate::error::*;
 use crate::file::{File, SeekWhence};
 
 impl File {
-    pub fn tell(&mut self) -> Result<usize> {
+    pub fn tell(&mut self) -> Result<u64> {
         return self.seek(SeekWhence::CurrentPos(0));
     }
 }
