+++ src/platform/apple/prelude/insert.rs
@@ -4,5 +4,5 @@ pub trait Insert<B>: Sized
 where
     B: AsRef<str>,
 {
-    fn insert(&mut self, whence: SeekWhence, buf: B) -> crate::error::Result<usize>;
+    fn insert(&mut self, whence: SeekWhence, buf: B) -> crate::error::Result<u64>;
 }
